Actualización Gigante
3 microservicios nuevos
Implementación Swagger
Implementación HATEOAS
Implementación API Gateway

+++ bicicleta-service/src/main/java/com/bicirepair/bicicleta_service/repository/BicicletaRepository.java
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface BicicletaRepository extends JpaRepository<Bicicleta, Integer> {
+public interface BicicletaRepository extends JpaRepository<Bicicleta, Long> {
     List<Bicicleta> findByMarca(String marca);
     List<Bicicleta> findByModelo(String modelo);
-    List<Bicicleta> findByIdCliente(int idCliente);
+    List<Bicicleta> findByIdCliente(Long idCliente);
 }
