+++ producto-service/src/main/java/com/bicirepair/producto_service/repository/ProductoRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ProductoRepository extends JpaRepository<Producto, Integer> {
-    List<Producto> findByIdCategoria(int idCategoria);
+public interface ProductoRepository extends JpaRepository<Producto, Long> {
+    List<Producto> findByIdCategoria(Long idCategoria);
     List<Producto> findByNombreProducto(String nombreProducto);
 }
